% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{bf_lowpass}
\alias{bf_lowpass}
\title{Lowpass or Highpass a timeseries}
\usage{
bf_lowpass(data, nm)
}
\arguments{
\item{data}{Timeseries}

\item{nm}{The cutoff frequency in units of time (i.e. use 12 with monthly data for 1-year frequency cutoff)}

\item{y}{timeseries to smooth}
}
\value{
The filtered timeseries
}
\description{
Lowpass or Highpass a timeseries using a Butter filter
}
\examples{
bf_lowpass (sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
bf_highpass(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
