% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dateincr}
\alias{dateincr}
\title{Dates increment}
\usage{
dateincr(ymd, incr = 1)
}
\arguments{
\item{ymd}{Date in YYYYMMDD format (either string or numeric)}

\item{incr}{Increment in days}
}
\value{
The new date
}
\description{
Increment dates by a certain number of days
}
\examples{
dateincr(20120110,10)
}
