% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{days_per_month}
\alias{days_per_month}
\title{Returns days of each month}
\usage{
days_per_month(year = NULL)
}
\arguments{
\item{year}{Optional (default non-leapfrog calendar)}
}
\value{
The vector of days for the 12 months
}
\description{
Returns days of each month, as a function of year
}
\examples{
days_per_month(2005)
}
