% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{demean}
\alias{demean}
\title{Remove the mean}
\usage{
demean(data)
}
\arguments{
\item{data}{Timeseries}
}
\value{
The anomaly timeseries
}
\description{
Remove the mean from data
}
\examples{
demean(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
