% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensrel.R
\name{ensrel}
\alias{ensrel}
\title{Compute Ensemble Reliability Budget}
\usage{
ensrel(obs, ens, obs_err, plot = TRUE, conf.l = 0.9)
}
\arguments{
\item{obs}{Vector of observed values (minimum 10 observations), of dimension n}

\item{ens}{Matrix containing the ensemble data at observation location, of dimension n x p (p being the ensemble size, minimum )}

\item{obs_err}{Value (constant) or vector of observation errors (either scalar or vector of dimension n)}

\item{plot}{Logical value for plotting the ensemble reliability budget as a barplot}

\item{conf.l}{Confidence level to test the residual}
}
\value{
A list with the ensemble reliability budget terms, including the p-value and the significance of the t-test for the null hypothesis (zero residual). A plot will be also produced (if plot == TRUE).
}
\description{
This function computes the ensemble reliability budget from observations
and observation-equivalent ensemble data.
}
\examples{
ensrel(rnorm(100,mean=25,sd=1),array(rnorm(600,mean=25.5,sd=1.0),dim=c(100,6)),0.2,conf.l=0.99)
}
