% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{get_att}
\alias{get_att}
\title{Get an attribute from NetCDF file}
\usage{
get_att(f, v, a)
}
\arguments{
\item{f}{Filename}

\item{v}{Variable name}

\item{a}{Attribute name}
}
\value{
The attribute
}
\description{
This function reads a netcdf attribute
for quick use
}
\examples{
temp_mv<-get_att("myfile.nc","temperature","missing_value")
}
