% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{get_fld}
\alias{get_fld}
\title{Get a field from NetCDF file}
\usage{
get_fld(f, v, start = NULL, count = NULL)
}
\arguments{
\item{f}{Filename}

\item{v}{Variable name}

\item{start}{Optional array, specifying the "start"}

\item{count}{Optional array, specifying the "count"}
}
\value{
The field
}
\description{
This function reads a netcdf variable
for quick use
}
\examples{
temp<-get_fld("myfile.nc","temperature")
}
