% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inserror}
\alias{inserror}
\title{Insitu observational errors}
\format{
A data frame with 3000 depths
}
\usage{
inserror
}
\description{
A dataset containing the in-situ observational errors for temperature
and salinity as a function of depth. Global values.
}
\keyword{datasets}
