% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{periodogram}
\alias{periodogram}
\title{Calculate periodogram}
\usage{
periodogram(data, plot = TRUE)
}
\arguments{
\item{data}{Timeseries}

\item{plot}{Plot the frequency space data (default to TRUE)}
}
\value{
The frequencies
}
\description{
Calculate (and plot) periodogram
}
\examples{
periodogram(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
