% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{put_fld}
\alias{put_fld}
\title{Put a field on an (existing) NetCDF file}
\usage{
put_fld(f, v, var)
}
\arguments{
\item{f}{Filename}

\item{v}{Variable name}

\item{var}{Field}
}
\description{
This function puts a netcdf variable
for quick use
}
\examples{
put_fld("myfile.nc","temperature",temp)
}
