% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankhist.R
\name{rankhist}
\alias{rankhist}
\title{Compute Rank Histogram}
\usage{
rankhist(obs, ens, plot = TRUE, conf.l = 0.9)
}
\arguments{
\item{obs}{Vector of observed values (minimum 10 observations), of dimension n}

\item{ens}{Matrix containing the ensemble data at observation location, of dimension n x p (p being the ensemble size, minimum )}

\item{plot}{Logical value for plotting the ensemble reliability budget as a barplot}

\item{conf.l}{Confidence level to test the residual}
}
\value{
A list with the ranks, including the RMSE of this histogram (w.r.t. to te perfect-ensemble value). A plot will be also produced (if plot == TRUE).
}
\description{
This function computes the ensemble rank histogram budget from observations
and observation-equivalent ensemble data.
}
\examples{
rankhist(rnorm(100,mean=25,sd=1),array(rnorm(600,mean=25.5,sd=1.0),dim=c(100,6)))
}
