% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{rm_seas}
\alias{rm_seas}
\title{Remove Seasonal signals}
\usage{
rm_seas(data, useclim = FALSE)
}
\arguments{
\item{data}{Timeseries}

\item{useclim}{Use monthly climatology instead of fitting to estimate the seasonal signal (Default to FALSE)}
}
\value{
The timeseries with seasonal signal removed
}
\description{
Remove annual and semi-annual signals from data
}
\examples{
rm_seas(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
