% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{rm_trend}
\alias{rm_trend}
\title{Remove Trend}
\usage{
rm_trend(data)
}
\arguments{
\item{data}{Timeseries}
}
\value{
The detrended timeseries
}
\description{
Remove trend from data
}
\examples{
rm_trend(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
