% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{rmse}
\alias{rmse}
\title{Compute the RMSE}
\usage{
rmse(y, x)
}
\arguments{
\item{y}{Verifying field}

\item{x}{Verified field}
}
\value{
RMSE Value
}
\description{
Compute the Root Mean Square Error
}
\examples{
rmse(rnorm(100,mean=3,sd=2),rnorm(100,mean=3.5,sd=2.5))
}
