% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{smooth}
\alias{smooth}
\title{Smooth a timeseries}
\usage{
smooth(x, y, sp = 0.2)
}
\arguments{
\item{x}{time data}

\item{y}{timeseries to smooth}

\item{sp}{Spar parameter of splines (default to 0.2)}
}
\value{
The smoothed timeseries
}
\description{
Calculate a smoothed version of the timeseries using splines
}
\examples{
smooth(sin( (1:240)*2*pi/12 )+rnorm(240,sd=0.3)+(1:240)*0.01)
}
